minor fixes

+++ src/network/dense.cpp
@@ -1,8 +1,8 @@
 #include "include/dense.hpp"
 
 // Constructor
-Dense::Dense(long input_size, long output_size){
-    this->weights = torch::randn({output_size, input_size}) / std::sqrt(static_cast<float>(input_size)); // W ~ N(0, 1/sqtr(n_in))
+Dense::Dense(int input_size, int output_size){
+    this->weights = torch::randn({output_size, input_size}) / std::sqrt(static_cast<double>(input_size)); // W ~ N(0, 1/sqtr(n_in))
     this->bias = torch::randn({output_size, 1}); // b ~ N(0, 1)
 }
 
@@ -24,5 +24,3 @@ torch::Tensor Dense::backward(torch::Tensor output_gradient, double eta) {
 
     return input_gradient;
 }
-
-
